[FEAT] #7 - '기획전'관련 repository 코드 추가
