+++ app/src/main/java/com/seom/banchan/data/repository/MenuRepositoryImpl.kt
@@ -1,10 +1,13 @@
 package com.seom.banchan.data.repository
 
-import com.seom.banchan.data.api.response.BestMenuResponse
+import com.seom.banchan.data.source.MenuDataSource
+import com.seom.banchan.domain.model.CategoryModel
 import com.seom.banchan.domain.repository.MenuRepository
 
-class MenuRepositoryImpl : MenuRepository {
-    override fun getBestMenus(): Result<BestMenuResponse> {
-        return Result.success(BestMenuResponse(emptyList()))
+class MenuRepositoryImpl(
+    private val menuDataSource: MenuDataSource
+) : MenuRepository {
+    override suspend fun getBestMenus(): Result<List<CategoryModel>> {
+        return menuDataSource.getBestMenus()
     }
 }
