+++ app/src/main/java/com/seom/banchan/di/CoroutineModule.kt
@@ -1,0 +1,14 @@
+package com.seom.banchan.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.Dispatchers
+
+@Module
+@InstallIn(SingletonComponent::class)
+object CoroutineModule {
+    @Provides
+    fun provideIODispatcher() = Dispatchers.IO
+}
