+++ app/src/main/java/com/seom/banchan/di/RepositoryModule.kt
@@ -1,0 +1,17 @@
+package com.seom.banchan.di
+
+import com.seom.banchan.data.repository.MenuRepositoryImpl
+import com.seom.banchan.domain.repository.MenuRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+    @Binds
+    abstract fun bindMenuRepository(
+        menuRepositoryImpl: MenuRepositoryImpl
+    ): MenuRepository
+}
