+++ app/src/main/java/com/seom/banchan/domain/repository/MenuRepository.kt
@@ -1,10 +1,11 @@
 package com.seom.banchan.domain.repository
 
 import com.seom.banchan.data.api.response.BestMenuResponse
+import com.seom.banchan.domain.model.CategoryModel
 
 interface MenuRepository {
     /**
      * 기획전 메뉴 정보 요청
      */
-    fun getBestMenus(): Result<BestMenuResponse>
+    suspend fun getBestMenus(): Result<List<CategoryModel>>
 }
